extracting some helper functions into utils folder
